Add documentation on index dates
These were added in this PR here: https://github.com/opensafely/cohort-extractor/pull/255

The main reason for adding them is to support extracting data for measures in the aftershocks project. It means that the same study definition can be run with the index date set to successive months in order to generate monthly timeseries data.