Explain the force run argument
This commit adds a single-element accordion (hidden, by default) that
explains the force run argument to the `opensafely run` command. I'm going
to assume that as it was helpful to me to write this explanation, then
it will be helpful to other people to read it. I feel the main
contributions are to highlight that when you specify an action, it is
always run; and that it's useful to think of all actions as dependencies
of the `run_all` action.