Tidy "Analysis workflow" page
* Add links.
* Clarify which steps to repeat in study development.

This might be the first documentation that someone new to OpenSAFELY
reads.

This document reads well as a quick overview. It's probably helpful for
someone's sense of "what do I need to read?" to directly link to
these concepts. A reader can then know that there's more information on
this topic, even if they don't read it there and then.

Also fix typos:

* Add missing "of".
* Fix "it's" -> "its".