Tidy "Using Git effectively" page
Clarify what a pull request does and change the subheading to relate to
the content and be clearer.

Also searching for `gitting "version control"` doesn't appear to show
many obvious uses as a verb itself. The most popular search results used
it as some pun on "getting" in relation to use of Git.