Add a first pass guide explaining Gitpod
After trying Gitpod, it seemed to be functional enough to allow a
non-technical user to start using OpenSAFELY.

This is a rough draft of documentation explaining that. There are a
couple of rough edges as marked by TODOs. It also needs proofing and the
links checking. Some screenshots would also be good to include.

It is not yet clear how this would best integrate into the current
documentation. This follows the Getting Started guide, but unnecessary
duplication is not good.

There is also a question over whether Gitpod change their free plan to
make it less useful for trying out OpenSAFELY. Though there is a
self-hosted version that is open source.