Use HTML headings in tabbed sections
The table of contents is not particularly customisable in mkdocs. The
problem with using tabbed sections for alternate paths is that headings
within those tabs end up in the table of contents. This means the table
of contents is both muddled and contains links that may not work,
because the relevant tab may not be visible.

The workaround was taken from this [mkdocs-material
discussion](https://github.com/squidfunk/mkdocs-material/discussions/2532).

Note that the question there is on restricting the depth of headings
that appear.  But we want something different: to suppress the headings
from being included in the table of contents if they are within tabs.
These headings are not necessarily at a given depth (although we could
constrain the writing of the document in this way).

This is not an ideal fix but is limited to this page. The alternative is
that we split this guide into pages corresponding to different paths,
but this lends itself to duplication.

This behaviour is documented with a comment, because the reasoning for
it may not be obvious.