Move details of Gitpod workspaces to own page
Away from Getting Started. It's useful background detail for people
who choose to use Gitpod for working with OpenSAFELY, but for those just
testing things out, it's a distraction. Those "trial" users shouldn't
ever encounter any of the limits: the demonstration takes a short time
to work through and it's, by default, in a public repository.