Fix toolchains' access to config
config is used by three toolchains: Cloudflare Pages (`.venv`), Data
Builder (`databuilder/.venv`), and pre-commit (the tool). To allow each
toolchain to access config, we ensure that flake8's per-file ignores are
relative to the project root, and we run the check and fix recipes in
databuilder/justfile from the project root.

You might be thinking "Cloudflare Pages doesn't need config; so why not
move config to Data Builder?" Well, then we'd break pre-commit (the
tool). You then might be thinking "Why not move pre-commit (the tool) to
Data Builder? Well, then everyone who cloned the repo would have to
setup both Cloudflare Pages and Data Builder to commit, even if the
commit was unrelated to Data Builder (see #619). As well as being
unnecessary, setting up both toolchains is a pain, because
`.python-version`, which is used by Dependabot, prevents pyenv from
putting 3.7 and 3.10 on the PATH.