Add first pass of notes for system integrators
And data providers.

This is a minimal starting point for dealing with
[this documentation issue](https://github.com/opensafely-core/backend-server/issues/61),
taking [@sebbacon's comment on GitHub](https://github.com/opensafely-core/backend-server/issues/61#issue-1223798579)
and rewriting.

The [content linked in the Google Doc](https://github.com/opensafely-core/backend-server/issues/61#issuecomment-1115864263)
is a good contender to replace this, but might require more of a review
and rewrite.

This commit at least makes some information known and hopefully reduces
the number of direct queries we get that could be answered with the
content here.