Use single quotes in scheduled workflow run
I mistakenly thought that these had to be single quotes from the
[documentation](https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#schedule).

In fact, the non-triggering of the scheduled run was a timezone issue as
discussed in this PR (#728).

However, a lot (thought not all) of the examples I've seen in GitHub's
documentation use single quotes, and I'm also sold by the arguments in
this [Stack Overflow discussion](https://stackoverflow.com/q/19109912)
that it's simplest to use single quotes by default when they are needed
— and they are needed here, because of the `*` — unless escape sequences
are needed.

Also, single quotes are used almost everywhere else in the workflows in
this repository. The only exception is the `lychee` arguments, but those
arguments contain single quotes, and therefore must be escaped. While
you can escape a single quote within a single quote YAML string, this
requires repeating the quote, making it more confusing for a human to
read.