Allow non-strict builds
In strict mode, MkDocs halts the build on a warning. griffe likes
warnings, especially about cohort-extractor's missing type annotations.
This is unhelpful, because we're not going to add type annotations to
cohort-extractor. However, we do want to build the documentation. So,
until we remember why we needed strict builds, and figure out how to
silence griffe, we'll allow non-strict builds.