Switch Python versions to be 3.9 everywhere
Not entirely sure why the Data Builder directory was 3.10.

Data Builder is currently [tested with
3.9](https://github.com/opensafely-core/databuilder/blob/08b3895f79d155075d2c295f5d64619da14003ad/.github/workflows/main.yml#L30), and the
opensafely-databuilder-mkdocs plugin is [tested with
3.7](https://github.com/opensafely-core/mkdocs-opensafely-databuilder/blob/7a1686a5f887e3b35d8dbe5fffd4bcfa066f0599/.github/workflows/main.yml#L30). (Which we
should possibly go and change.)

In principle, we could move to Python 3.10 too. The reason for not doing
so is that Data Builder is only tested with 3.9.

If we upgrade Data Builder, then we could do the same here.