Install cohort-extractor as a submodule
This avoids installing into the same virtualenv as the documentation and
allows us to:

* more easily unify the Python version we're using for
  building the documentation and generating snippets.
* choose a newer version of Python than cohort-extractor uses

All we actually need are the docstrings right now. If we later want to
test cohort-extractor snippets, we might have to revisit this.

Note that the whitespace changes are because pre-commit is now running
on the entire repository, not just the `databuilder/` subdirectory.