Update Data Builder version
This updates to a version which features the extensive changes to how
backends and contracts are implemented: opensafely-core/databuilder/pull/680

The specific version currently used is on a branch; see the PR:
opensafely-core/databuilder/pull/703.

The tutorial dataset definitions have been amended to work as is.

Adding a tutorial backend has two benefits:

* We're not referring to real backends in the tutorial.
* We can arbitrarily add the tables that we want, even if they are not
  enabled for any real backend.