Add `project.yaml` explainer for Data Builder
This is not completely finished. However, to do so, requires getting the
`project.yaml` working, which may not be possible yet.

In future, we may prefer to integrate this information earlier, and
replace the existing examples that run with Data Builder directly, with
entire OpenSAFELY projects: one for each dataset definition.

For now, this at least makes the connection between Data Builder and
OpenSAFELY `project.yaml` explicit.