Clarify current installation state
@milanwiedemann tried working through the installation as written and
found it troublesome.

We managed to get it working, but it would have been helpful to document
what you should do right now.

Docker doesn't work at all as written, because you need to make the
local files accessible to the Docker container somehow (via volumes or
bind mounts).

We intend to move to the OpenSAFELY CLI to run Data Builder. This may
not quite work yet, although there has been work on it. That
functionality needs reviewing again.

The Python route is not what we want people to use by default, but is
the currently working option.