Pin the Docker image used for the examples
For now until a better solution is implemented.

We want the Python installed package in the requirements, and this
Docker image to be in sync. Ideally, we would just always use the latest
release of the current major version, but we can't easily pin this in
the Python requirements because we need an actual hash for `pip`.

It may be that we later:

1. Add a `just` recipe to update the code.
2. Add an automated check to CI to do so.