Mention `opensafely exec` to run examples
This may need some more polish, so keep both options in for now.
We may later keep the use of `project.yaml` towards the end of the
tutorial for simplicity.

But `opensafely exec` has advantages of:

* being very similar to running `databuilder` via a `pip`-installed
  package
* removing the need for `project.yaml` immediately, which is a
  simplification when getting started

I had issues running it, so I've not done anything like drop the
existing `project.yaml` that covers all dataset definitions.

If we do drop the `project.yaml` for routine running of the tutorial,
we can return to including a simple `project.yaml` just for the minimal
dataset definition. This makes the tutorial exercises for explaining
`project.yaml` more useful.