Workaround issue by using a non-relative URL
The paths in the include are relative to the file where it is included.

We could fix this to correct the relative path.

But, because we've reorganised the file structure, we have different
Markdown files at different levels in the directory hierarchy now including
this same include file.

Therefore, there is no single relative path that will fix the link in
all places that the file is included: we might be including this file
from Markdown source files that are either two or three levels deep in
the directory hierarchy.

As far as I can tell, there is no direct solution for this, and the
simplest fix is to use an absolute URL that is not part of the docs at
all:

https://github.com/mkdocs/mkdocs/issues/2404