Reinstate scheduled docs link check every Tuesday
We tried using a Slack reminder, but you can't control which specific
weekdays a Slack reminder appears, if the reminders aren't weekly.

(You can schedule reminders like "on the 15th day of every month", or
"every Tuesday" for example, but not "every fourth Tuesday".)

This is fairly inflexible and ended up being posted on a Sunday. Which
no-one spotted.

Instead, go back to running the job automatically, but reduce the
frequency to once per week on a given weekday, which seems reasonable.
Previously, the job ran automatically once per day, when removed, which
would be irritating if no-one is dealing with any persistent issue
immediately.

Running automatically also removes the need for someone to first read
the reminder, then manually run the workflow and *then* have the error
appear in Slack.