Move imported documentation for link checking
With this documentation in `imported_docs`, the surrounding
documentation cannot find the imported documentation in the correct
relative path. This is fine for building the site because the multirepo
plugin handles this.

But the Markdown links are relative, and the link checker has no
knowledge of this, so let's move the files to the correct place for
checking.

We could alternatively actually store the imported documentation in
`docs/` in the multirepo plugin configuration instead of
`imported_docs/`.

However, I have not done this for now because:

* the directory used for importing documentation is considered a temporary
  directory by the multirepo plugin (although it is possible to get this
  directory to persist via configuration)
* this move is only required for this check, so having a workaround for
  the check seems reasonable, instead of changing the configuration