Ignore generated docs snippets in link check
This is a workaround.

We actually do want to check these for broken links.

However, in the current state, it is not easily possible, and the check
will always fail even if the links in those generated docs are valid.

Note that this problem is caused by the file being snippets, rather than
generated.

In more detail:

* the generated docs are stored in their own directory
* the generated docs use relative paths to, for example, refer to the
  parent directory
* the generated docs are then included verbatim in other documentation
  files via the [snippet syntax](https://facelessuser.github.io/pymdown-extensions/extensions/snippets/)
  in the Python Markdown extensions
* those relative paths then become relative to *where the file is
  included* and can be correctly resolved
* when lychee checks the files, it has no knowledge of the inclusion and
  incorrectly tries to resolve the path to, for example, the parent of
  the generated docs directory. We actually want it to check the parent
  of the directory containing the file that includes the snippet.

If desired, solving this could be via one of:

* keep the generated docs as they are being included
* use absolute paths in the included docs, though that isn't really
  supported by MkDocs (but might be in future?)
  https://github.com/mkdocs/mkdocs/issues/1592
* write a MkDocs hook or shell script that copies the included files to
  the correct place for relative links to work, on build for the link
  check only
* checking the HTML for links, not Markdown except that I tried this as
  the HTML checks started checking 12,000 URLs, not 600, and lychee
  can result in a single URL being requested multiple times in different
  worker threads

Snippets can, in principle, be included recursively, so this is
potentially more complicated to solve than you might expect at first
glance.

lychee does have a remap URL feature, but that only applies globally, not
per-file.