Make `docs.opensafely.org` URLs point to Markdown
These links will take you to a different site entirely, when you access
a version of the site that is not on the `docs.opensafely.org` domain.

This can give rise to unexpected behaviour.

For instance, let's say a page is deleted in your local version of the
site, but not on `docs.opensafely.org` and that it is linked from in a
page on the site. The link from your local version would still work in
preview, even though it would not when your version of the site is
published.

It's possible then to click through from a local version preview to the
live docs site, and then end up on the live site, continue clicking
around and not notice.