Add ehrQL Pygments configuration
We now use `ehrql` syntax labels for complete dataset definitions in the
ehrQL documentation: opensafely-core/ehrql#1648.

This documentation repository needs to know about this configuration
too. This duplication is a little bit irritating, but it is because our
multi-repository configuration doesn't have any kind of configuration
inheritance or reuse. See opensafely-core/ehrql#1126

A future improvement might be to use the Material for MkDocs projects
plugin once it's available in the free version which might also fix this
duplication: #1317.