Add "how-to" guides to table of contents
This does not add the "setting up OpenSAFELY" guides listed in the
Getting Started how-to section.

This is because it's not possible to link to anchors from within
`mkdocs.yml` which limits the utility. However, this is also a
reasonable approach given that we want to promote Codespaces as the
suggested development environment for researchers.

In future, we may wish to restructure these anyway, to perhaps have a
single page of guidance per operating system, instead of a page for each
piece of software to be installed.

Also note that some of these pages listed in this part of the table of
contents are not in the corresponding file hierarchy in the repository,
and are at the top-level. We can leave these for now pending possible
future restructuring. If the pages remain as is, they should eventually
be moved into `docs/getting-started/how-to/`.