Remove RStudio login instructions
RStudio server is configured to run without needing authentication as of [this](https://github.com/opensafely-core/research-template-docker/commit/cdd84e6e5fcce8eb95826882da6e170f7e69ab68) commit. This change removes the instruction to log in to RStudio as the user is no longer prompted with a login screen.