Fix "Commit and push" section
The "Commit and push" section was missing an "add" step. Thanks for
spotting this, @alschaffer.

It's not possible to stage and commit untracked files with a single
command (i.e. `git commit --all` only stages files that were tracked),
so we add an "add" step rather than update the "commit" step.