Describe how to handle `.gitignore`
The "How to add GitHub Codespaces to your project" section instructs a
user to download files from opensafely/research-template and copy them
to their project. One file in this folder is a dotfile: `.gitignore`.
However, the "Download raw file" button removes the dot.

Here, we work around this understandable/unhelpful behaviour by asking
the user to rename `gitignore` to `.gitignore`. We deliberately avoid
describing *how* to do so, because operating systems, tools, and set-ups
vary.

Closes #1533