Add bare minimum instructions for adding a paper
The full instructions have been moved to the
internal team manual.

Google will still return search results for the
old public-facing instructions at this URL.

To avoid 404 errors, this page is recreated and
points anyone who ends up there to the internal docs.

This page is not added back into the menu as we
do not believe it is neccesary for non-Bennett
users to find it.