Notify Slack on failed scheduled link checks only
This avoids manually triggered checks sending noise to Slack, which is
particularly annoying if fixing the link check requires multiple runs in
doing so.

This also explicitly marks the `if` condition as an expression, which is
the default behaviour for `if`, but marking as such makes it clearer
that it is an expression.