update instructions for providing info on published papers
We need external opensafely users to provide info on all papers they have pre-printed or published.

The intention here is to provide a downloadable link to a template file, which contains clear instructions on what content to provide in the given format. I'd like the content of the template is visible on this page, so users can see what they need to provide without downloading the file). We could duplicate this info as formatted text on the page, but if anything changes then there are two things to update.

This commit does this by providing the template file in a code-block -- i'm hoping that it's possible to do some magic that makes this block downloadable as a .md file (or .txt file, if easer for users to work with).

Alternatively, we can provide a link to the template for download, and _also_ display the contents of the template on this page.

if neither are possible, the a downloadable link to the template and in-page content that described what is needed will suffice