Add minimal dev container configuration
For editing the site using a dev container locally, or in Codespaces
online.

The `deadsnakes` configuration is taken from the
`research-template-docker` repository:

https://github.com/opensafely-core/research-template-docker/blob/669a76d79c198925d788dedc34fe3c5a831872b5/Dockerfile#L20-L21

which avoids installing `software-properties-common` first (to get the
`add-apt-repository` command.

The dev container settings are based on the Bennett site configuration,
with a longer autosave delay, because the site currently takes a while
to build (#1397). Arguably, the save delay could be longer still.

No Markdown plugins are installed yet; maybe that's something we could
do in future.