Extend Codespaces retention period
We've changed the Codespaces retention period from 14 days to the maximum
allowed time of 30 days. This is to reduce the chances of researchers losing
data when their Codespace is automatically deleted.

The notice informing users of this change has been moved further up the page,
so they're more likely to see it.