Remove note on cohort-extractor docstrings
As spotted by @alschaffer in this issue:

https://github.com/opensafely/documentation/issues/1573#issuecomment-2393631860

The relevant file has been edited just four times in the last two years,
and not since mid-2023.

Realistically, the only people who might make small updates to
cohort-extractor are likely to be developers now, and this note probably
just mystifies anyone else who encounters it.