Tweak Codespace retention warning
We've removed the explicit policy that enforces the 30 day limit, which means
that people can opt-out of auto-deletion on a per-Codespace basis. I don't
expect this to happen very often and I don't think a warning box is the right
place to explain how to do it, so I've avoided going into any detail here.