Remove outdated warning
It used to be the case that case-control studies were only supported by the TPP backend,
since the `patients_with_value_from_file` feature in cohortextractor was only available there.
We therefore had to warn users about this.

Now that we have switched over to ehrQL and the `table_from_file` feature,
we can support case-control studies in all backends.