Update Codespaces billing information
This information was out of date, as we changed the policy a while ago to allow
any approved OpenSAFELY users to use GitHub Codespaces with any repo in the
`opensafely` organisation. OpenSAFELY users are less likely to use Codespaces
with their own account, except for the getting started tutorials, so the
billing information is less relevant than it was but still worth keeping.

Fixes #1788