Remove "dataset" and "backend" from glossary
The term "dataset" is used in a variety of different senses within
OpenSAFELY and its inclusion here was a regular source of confusion.

While "backend" wasn't quite as problematic it is still used in multiple
senses within the docs and is a sufficiently general term that we don't
want to define it here.